Automated_Suite/requirements.txt
apscheduler==3.9.0
beautifulsoup4==4.11.1
feedparser==6.0.10
nltk==3.7
numpy==1.22.4
pyautogui==0.9.53
pytweepy==4.10.1
requests==2.28.1
scipy==1.8.1
smtplib==3.11.1